Photo Tagging Functionality
Photos can now be associated with a group, and/or be displayed
publically or privately.

When a user decides to upload a photo, they are now taken to a new
screen. This new screen displays the image, along with options the user
has for the photo. They can add it to a group, and make it a private
photo so it is only viewable by friends.