Rating System Added
There is now an initial rating system when the user tries to upload a
photo.

The user is asked to rate the photo when they upload it, from 1-5 stars.
This rating is then used to order the photos in a group list from best
to worst, so the better photos are seen first.