Create Pin Location Working
Users can now select a location on the map if the photo has no GPS
information. This new pin is then associated with the image and can be
uploaded to the map.

A new 'location preview' window has also been added to the upload
screen. At the minute, it simply shows a map centred on where the user
is, but functionality to display the currently selected photo's location
will be added soon.